# specs/001-country-service/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Country WebAPI Service
  description: |
    High-performance, read-optimized Country WebAPI Service serving as the canonical source for country data
    across the Maliev microservices ecosystem. Features multi-layer caching, ETag-based conditional requests,
    and comprehensive administrative endpoints with optimistic concurrency control.
  version: 1.0.0
  contact:
    name: Maliev API Support
    email: api-support@maliev.com
servers:
  - url: https://api.maliev.com/country/v1
    description: Production
  - url: https://api-staging.maliev.com/country/v1
    description: Staging
  - url: https://api-dev.maliev.com/country/v1
    description: Development
  - url: http://localhost:5000/country/v1
    description: Local Development
...
paths:
...
  /countries/{id}:
    get:
      tags: [Countries]
      summary: Get country by ID
      description: Returns a single country by database ID. Cached with 15-minute TTL.
      operationId: getCountryById
      security: []
      parameters:
        - $ref: '#/components/parameters/CountryId'
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Successful response
          headers:
            ETag:
              schema:
                type: string
              description: Entity version for optimistic concurrency
            Last-Modified:
              schema:
                type: string
                format: date-time
              description: Last modification timestamp
            X-Cache:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountryResponse'
              examples:
                success:
                  $ref: '#/components/examples/CountryDetailSuccess'
        '304':
          description: Not Modified (ETag match)
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
  /admin/countries/{id}:
    put:
      tags: [Admin]
      summary: Update country (full replacement)
      description: |
        Updates all fields of an existing country. Requires If-Match header with current ETag
        for optimistic concurrency control.
      operationId: updateCountry
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CountryId'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCountryRequest'
      responses:
        '200':
          description: Country updated successfully
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Conflict - version mismatch or duplicate ISO code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '412':
          description: Precondition Failed - If-Match header doesn't match current ETag
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/TooManyRequests'

    patch:
      tags: [Admin]
      summary: Update country (partial update)
      description: Updates only specified fields. Requires If-Match header.
      operationId: patchCountry
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CountryId'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchCountryRequest'
      responses:
        '200':
          description: Country updated successfully
          headers:
            ETag:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountryResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '412':
          $ref: '#/components/responses/PreconditionFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'

    delete:
      tags: [Admin]
      summary: Delete country (soft delete)
      description: Soft-deletes country by setting isActive=false. Requires CountryAdmin role.
      operationId: deleteCountry
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CountryId'
      responses:
        '204':
          description: Country deleted successfully (no content)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /admin/countries/{id}/hard-delete:
    delete:
      tags: [Admin]
      summary: Delete country permanently (hard delete)
      description: |
        Permanently deletes country from database. Requires SuperAdmin role.
        WARNING: This operation is irreversible.
      operationId: hardDeleteCountry
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CountryId'
      responses:
        '204':
          description: Country permanently deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'

  # ===== BULK IMPORT ENDPOINTS (Authenticated) =====
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Maliev Auth Service
  parameters:
    CountryId:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
      example: 123
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
      description: ETag from previous GET request (required for optimistic concurrency)
      example: '"abc123def456"'
    IfNoneMatch:
      name: If-None-Match
      in: header
      schema:
        type: string
      description: ETag from previous GET request (for conditional GET)
      example: '"abc123def456"'
  schemas:
    CountryResponse:
      type: object
      required: [id, iso2, iso3, name, isActive, version, createdAtUtc, lastModifiedUtc]
      properties:
        id:
          type: integer
          format: int64
          example: 123
        iso2:
          type: string
          pattern: '^[A-Z]{2}$'
          example: US
        iso3:
          type: string
          pattern: '^[A-Z]{3}$'
          example: USA
        name:
          type: string
          maxLength: 100
          example: United States
        officialName:
          type: string
          maxLength: 200
          nullable: true
          example: United States of America
        numericCode:
          type: string
          pattern: '^[0-9]{3}$'
          nullable: true
          example: "840"
        capital:
          type: string
          maxLength: 100
          nullable: true
          example: Washington D.C.
        region:
          type: string
          maxLength: 50
          nullable: true
          example: Americas
        subregion:
          type: string
          maxLength: 50
          nullable: true
          example: Northern America
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
          nullable: true
          example: 38.8951
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
          nullable: true
          example: -77.0364
        demonym:
          type: string
          maxLength: 50
          nullable: true
          example: American
        areaKm2:
          type: number
          format: double
          minimum: 0
          nullable: true
          example: 9833520
        population:
          type: integer
          format: int64
          minimum: 0
          nullable: true
          example: 331900000
        giniCoefficient:
          type: number
          format: double
          minimum: 0
          maximum: 100
          nullable: true
          example: 41.4
        timezones:
          type: array
          items:
            type: string
          example: ["UTC-10:00", "UTC-05:00", "UTC-04:00"]
        borders:
          type: array
          items:
            type: string
            pattern: '^[A-Z]{3}$'
          example: ["CAN", "MEX"]
        callingCodes:
          type: array
          items:
            type: string
          example: ["+1"]
        topLevelDomains:
          type: array
          items:
            type: string
          example: [".us"]
        currencies:
          type: object
          additionalProperties:
            type: object
            properties:
              name:
                type: string
              symbol:
                type: string
          example:
            USD:
              name: United States dollar
              symbol: $
        languages:
          type: object
          additionalProperties:
            type: string
          example:
            eng: English
        translations:
          type: object
          additionalProperties:
            type: string
          example:
            spa: Estados Unidos
            fra: États-Unis
        flags:
          type: object
          properties:
            svg:
              type: string
              format: uri
            png:
              type: string
              format: uri
        coatOfArms:
          type: object
          nullable: true
          properties:
            svg:
              type: string
              format: uri
            png:
              type: string
              format: uri
        independent:
          type: boolean
          example: true
        unMember:
          type: boolean
          example: true
        landlocked:
          type: boolean
          example: false
        isActive:
          type: boolean
          example: true
        version:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        createdAtUtc:
          type: string
          format: date-time
          example: "2025-01-01T00:00:00Z"
        lastModifiedUtc:
          type: string
          format: date-time
          example: "2025-01-15T10:30:00Z"
    CreateCountryRequest:
      type: object
      required: [iso2, iso3, name]
      properties:
        iso2:
          type: string
          pattern: '^[A-Z]{2}$'
        iso3:
          type: string
          pattern: '^[A-Z]{3}$'
        name:
          type: string
          minLength: 1
          maxLength: 100
        officialName:
          type: string
          maxLength: 200
        numericCode:
          type: string
          pattern: '^[0-9]{3}$'
        capital:
          type: string
          maxLength: 100
        region:
          type: string
          maxLength: 50
        subregion:
          type: string
          maxLength: 50
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
        demonym:
          type: string
          maxLength: 50
        areaKm2:
          type: number
          minimum: 0
        population:
          type: integer
          format: int64
          minimum: 0
        giniCoefficient:
          type: number
          minimum: 0
          maximum: 100
        timezones:
          type: array
          items:
            type: string
        borders:
          type: array
          items:
            type: string
            pattern: '^[A-Z]{3}$'
        callingCodes:
          type: array
          items:
            type: string
        topLevelDomains:
          type: array
          items:
            type: string
        currencies:
          type: object
        languages:
          type: object
        translations:
          type: object
        flags:
          type: object
        coatOfArms:
          type: object
        independent:
          type: boolean
        unMember:
          type: boolean
        landlocked:
          type: boolean
    UpdateCountryRequest:
      allOf:
        - $ref: '#/components/schemas/CreateCountryRequest'
    PatchCountryRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        officialName:
          type: string
          maxLength: 200
          nullable: true
        capital:
          type: string
          maxLength: 100
          nullable: true
        region:
          type: string
          maxLength: 50
          nullable: true
        subregion:
          type: string
          maxLength: 50
          nullable: true
        population:
          type: integer
          format: int64
          minimum: 0
          nullable: true
        # ... (other patchable fields)
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          example: ValidationFailed
        message:
          type: string
          example: One or more validation errors occurred
        details:
          type: object
          nullable: true
        correlationId:
          type: string
          format: uuid
          nullable: true
  responses:
    BadRequest:
      description: Bad Request - validation errors
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not Found - resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    PreconditionFailed:
      description: Precondition Failed - If-Match header doesn't match current ETag
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    TooManyRequests:
      description: Too Many Requests - rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until rate limit resets
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  examples:
...
    CountryDetailSuccess:
      summary: Complete country detail
      value:
        id: 1
        iso2: US
        iso3: USA
        name: United States
        officialName: United States of America
        numericCode: "840"
        capital: Washington D.C.
        region: Americas
        subregion: Northern America
        latitude: 38.8951
        longitude: -77.0364
        demonym: American
        areaKm2: 9833520
        population: 331900000
        giniCoefficient: 41.4
        timezones: ["UTC-10:00", "UTC-05:00"]
        borders: ["CAN", "MEX"]
        callingCodes: ["+1"]
        topLevelDomains: [".us"]
        currencies:
          USD:
            name: United States dollar
            symbol: $
        languages:
          eng: English
        translations:
          spa: Estados Unidos
        flags:
          svg: https://flagcdn.com/us.svg
          png: https://flagcdn.com/w320/us.png
        independent: true
        unMember: true
        landlocked: false
        isActive: true
        version: "123e4567-e89b-12d3-a456-426614174000"
        createdAtUtc: "2025-01-01T00:00:00Z"
        lastModifiedUtc: "2025-01-15T10:30:00Z"
...